+++ Runtime/Dialogue/DialogueSequence.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace GameKit.Dialogue
+{
+    [CreateAssetMenu(fileName = "DialogueSequence", menuName = "GameKit/Dialogue/Dialogue Sequence")]
+    public sealed class DialogueSequence : ScriptableObject
+    {
+        [Tooltip("Dialogue entries played in order. Entries may be voiced or non-voiced.")]
+        public List<DialogueEntry> entries = new();
+    }
+}